Appended TabBarCoordinator
