--- p2p_wallet/Scenes/Main/Main.swift
@@ -1,10 +1,0 @@
-//
-//  Main.swift
-//  p2p_wallet
-//
-//  Created by Chung Tran on 26/04/2022.
-//
-
-import Foundation
-
-enum Main {}
